# swagger.yaml
swagger: "2.0"
info:
  title: API de Eventos
  description: API para cadastro de usuários, login e gerenciamento de eventos.
  version: "1.0.0"
servers:
  - url: http://localhost:5000

paths:
  /login:
    post:
      summary: Autentica o usuário
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                senha:
                  type: string
      responses:
        "200":
          description: Login bem-sucedido
        "401":
          description: Usuário ou senha inválidos

  /logout:
    post:
      summary: Faz logout do usuário
      responses:
        "200":
          description: Logout efetuado

  /cadastro:
    post:
      summary: Cadastra novo usuário
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                nome:
                  type: string
                login:
                  type: string
                senha:
                  type: string
      responses:
        "201":
          description: Usuário cadastrado com sucesso
        "409":
          description: Usuário já existe

  /users:
    get:
      summary: Lista os usuários cadastrados
      responses:
        "200":
          description: Lista de usuários
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: integer
                    nome:
                      type: string

  /eventos:
    get:
      summary: Lista todos os eventos
      responses:
        "200":
          description: Lista de eventos
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: integer
                    nome:
                      type: string
                    categoria:
                      type: string
                    data:
                      type: string
                    status:
                      type: string

    post:
      summary: Cria um novo evento
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                evento:
                  type: object
                  properties:
                    nome:
                      type: string
                    categoria:
                      type: string
                    data:
                      type: string
                    status:
                      type: string
      responses:
        "201":
          description: Evento criado com sucesso
        "400":
          description: Dados inválidos
        "500":
          description: Erro no servidor

  /eventos/{id}:
    delete:
      summary: Deleta um evento por ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Evento deletado com sucesso
        "404":
          description: Evento não encontrado
        "500":
          description: Erro no servidor